refactor: fieldStyles를 inputFieldStyles로 rename
