+++ src/feature/CardRegister/components/InfoInputSection/CvcField/CvCField.tsx
@@ -1,6 +1,6 @@
 import Input from '../../../../../common/components/Input/Input';
 import styled from 'styled-components';
-import {StyledField, ErrMessage} from '../fieldStyles';
+import {StyledField, ErrMessage} from '../inputFieldStyles';
 type Props = {
   cvcNumber: string;
   firstErrIdx: number;
