+++ src/feature/CardRegister/components/InfoInputSection/inputFieldStyles.ts
